fix: cast nonce sequence number to u64 to prevent wasm32 panic
seq[4..] requires 8 bytes, but usize::to_be_bytes() returns only 4
bytes on wasm32, causing copy_from_slice to panic at runtime.
Casting to u64 ensures 8-byte output on all platforms.

+++ ece/src/lib.rs
@@ -133,7 +133,7 @@ pub fn encrypt<IKM: AsRef<[u8]>, KI: AsRef<[u8]>, R: Iterator<Item = Vec<u8>>>(
 
     let records = records.enumerate().map(|(n, record)| {
         let mut seq = [0u8; 12];
-        seq[4..].copy_from_slice(&n.to_be_bytes());
+        seq[4..].copy_from_slice(&(n as u64).to_be_bytes());
         let key = derive_key(salt, ikm.as_ref());
         let nonce = derive_nonce(salt, ikm.as_ref(), seq);
         (key, nonce, record)
@@ -213,7 +213,7 @@ pub fn decrypt<IKM: AsRef<[u8]>>(
         .enumerate()
         .map(|(n, record)| {
             let mut seq = [0u8; 12];
-            seq[4..].copy_from_slice(&n.to_be_bytes());
+            seq[4..].copy_from_slice(&(n as u64).to_be_bytes());
             let key = derive_key(salt, ikm.as_ref());
             let nonce = derive_nonce(salt, ikm.as_ref(), seq);
             (key, nonce, record)
